+++ src/tsserver/_namespaces/ts.server.ts
@@ -5,4 +5,4 @@ export * from "../../server/_namespaces/ts.server";
 export * from "../../webServer/_namespaces/ts.server";
 export * from "../nodeServer";
 export * from "../webServer";
-export * from "../server";
+export * from "../common";
